requirements.txt
asgiref==3.8.1
asttokens==2.4.1
bcrypt==4.2.1
blinker==1.8.2
cachelib==0.13.0
cffi==1.16.0
click==8.1.7
colorama==0.4.6
comm==0.2.0
contourpy==1.2.0
cryptography==42.0.5
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
Django==5.1.1
dnspython==2.7.0
email_validator==2.2.0
executing==2.0.1
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
fonttools==4.46.0
greenlet==3.1.1
idna==3.10
ipykernel==6.27.1
ipython==8.18.1
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jupyter_client==8.6.0
jupyter_core==5.5.0
kiwisolver==1.4.5
MarkupSafe==3.0.1
matplotlib==3.8.2
matplotlib-inline==0.1.6
msgspec==0.19.0
nest-asyncio==1.5.8
numpy==1.26.2
packaging==23.2
parso==0.8.3
Pillow==10.1.0
platformdirs==4.0.0
prompt-toolkit==3.0.41
psutil==5.9.6
pure-eval==0.2.2
pycparser==2.21
Pygments==2.17.2
pyparsing==3.1.1
PyPDF2==3.0.1
python-dateutil==2.8.2
pywin32==306
pyzmq==25.1.1
six==1.16.0
SQLAlchemy==2.0.37
sqlparse==0.5.1
stack-data==0.6.3
tornado==6.4
traitlets==5.14.0
typing_extensions==4.12.2
tzdata==2024.1
wcwidth==0.2.12
Werkzeug==3.0.4
WTForms==3.2.1
